+++ crates/ruma-common/src/push/predefined.rs
@@ -1,8 +1,6 @@
 ///! Constructors for [predefined push rules].
 ///!
 ///! [predefined push rules]: https://spec.matrix.org/v1.2/client-server-api/#predefined-rules
-use indexmap::indexset;
-
 use super::{
     Action::*, ConditionalPushRule, PatternedPushRule, PushCondition::*, RoomMemberCountIs,
     Ruleset, Tweak,
@@ -20,48 +18,37 @@ impl Ruleset {
     ///   user's ID (for instance those to send notifications when they are mentioned).
     pub fn server_default(user_id: &UserId) -> Self {
         Self {
-            content: indexset![PatternedPushRule::contains_user_name(user_id)],
-            #[cfg(feature = "unstable-msc2677")]
-            override_: indexset![
+            content: [PatternedPushRule::contains_user_name(user_id)].into(),
+            override_: [
                 ConditionalPushRule::master(),
                 ConditionalPushRule::suppress_notices(),
                 ConditionalPushRule::invite_for_me(user_id),
                 ConditionalPushRule::member_event(),
                 ConditionalPushRule::contains_display_name(),
                 ConditionalPushRule::tombstone(),
+                #[cfg(feature = "unstable-msc3786")]
+                ConditionalPushRule::server_acl(),
                 ConditionalPushRule::roomnotif(),
+                #[cfg(feature = "unstable-msc2677")]
                 ConditionalPushRule::reaction(),
-            ],
-            #[cfg(not(feature = "unstable-msc2677"))]
-            override_: indexset![
-                ConditionalPushRule::master(),
-                ConditionalPushRule::suppress_notices(),
-                ConditionalPushRule::invite_for_me(user_id),
-                ConditionalPushRule::member_event(),
-                ConditionalPushRule::contains_display_name(),
-                ConditionalPushRule::tombstone(),
-                ConditionalPushRule::roomnotif(),
-            ],
-            #[cfg(feature = "unstable-msc3381")]
-            underride: indexset![
+            ]
+            .into(),
+            underride: [
                 ConditionalPushRule::call(),
                 ConditionalPushRule::encrypted_room_one_to_one(),
                 ConditionalPushRule::room_one_to_one(),
                 ConditionalPushRule::message(),
                 ConditionalPushRule::encrypted(),
+                #[cfg(feature = "unstable-msc3381")]
                 ConditionalPushRule::poll_start_one_to_one(),
+                #[cfg(feature = "unstable-msc3381")]
                 ConditionalPushRule::poll_start(),
+                #[cfg(feature = "unstable-msc3381")]
                 ConditionalPushRule::poll_end_one_to_one(),
+                #[cfg(feature = "unstable-msc3381")]
                 ConditionalPushRule::poll_end(),
-            ],
-            #[cfg(not(feature = "unstable-msc3381"))]
-            underride: indexset![
-                ConditionalPushRule::call(),
-                ConditionalPushRule::encrypted_room_one_to_one(),
-                ConditionalPushRule::room_one_to_one(),
-                ConditionalPushRule::message(),
-                ConditionalPushRule::encrypted(),
-            ],
+            ]
+            .into(),
             ..Default::default()
         }
     }
@@ -182,6 +169,23 @@ impl ConditionalPushRule {
             enabled: true,
             rule_id: ".m.rule.reaction".into(),
             conditions: vec![EventMatch { key: "type".into(), pattern: "m.reaction".into() }],
+        }
+    }
+
+    /// Matches [room server ACLs].
+    ///
+    /// [room server ACLs]: https://spec.matrix.org/v1.3/client-server-api/#server-access-control-lists-acls-for-rooms
+    #[cfg(feature = "unstable-msc3786")]
+    pub fn server_acl() -> Self {
+        Self {
+            actions: vec![],
+            default: true,
+            enabled: true,
+            rule_id: ".org.matrix.msc3786.rule.room.server_acl".into(),
+            conditions: vec![
+                EventMatch { key: "type".into(), pattern: "m.room.server_acl".into() },
+                EventMatch { key: "state_key".into(), pattern: "".into() },
+            ],
         }
     }
 }