+++ crates/ruma-macros/src/lib.rs
@@ -358,12 +358,8 @@ pub fn fake_derive_serde(_input: TokenStream) -> TokenStream {
     TokenStream::new()
 }
 
-/// A procedural macro for easily generating API endpoints.
-///
-/// Note that for technical reasons, the `ruma_api!` macro is only documented in [ruma-common], not
-/// here.
-///
-/// [ruma-common]: https://github.com/ruma/ruma/tree/main/ruma-common
+/// > ⚠ If this is the only documentation you see, please navigate to the docs for
+/// > `ruma_common::api::ruma_api`, where actual documentation can be found.
 #[proc_macro]
 pub fn ruma_api(input: TokenStream) -> TokenStream {
     let api = parse_macro_input!(input as Api);