+++ crates/ruma-common/CHANGELOG.md
@@ -23,7 +23,9 @@ Breaking changes:
 * Move `CanonicalJson`, `CanonicalJsonObject` and `CanonicalJsonError` out of
   the `serde` module and behind the cargo feature flag `canonical-json`
 * Make identifiers matrix URI constructors generic over owned parameters
+  * Split `RoomId` matrix URI constructors between methods with and without routing
 * Allow to add routing servers to `RoomId::matrix_to_event_uri()`
+* Move `receipt::ReceiptType` to `events::receipt`
 
 [spec]: https://github.com/matrix-org/matrix-spec-proposals/pull/3669
 
@@ -44,6 +46,8 @@ Improvements:
 * Add methods to sanitize messages according to the spec behind the `unstable-sanitize` feature
   * Can also remove rich reply fallbacks
 * Implement `From<Owned*Id>` for `identifiers::matrix_uri::MatrixId`
+* Add unstable default push rule to ignore room server ACLs events (MSC3786)
+* Add unstable support for private read receipts (MSC2285)
 
 # 0.9.2